[Gold IV] Title: 단어 수학, Time: 108 ms, Memory: 14220 KB -BaekjoonHub

+++ 백준/Gold/1339. 단어 수학/단어 수학.java
@@ -12,7 +12,7 @@ public class Main {
         for (int i = 0; i < N; i++) {
             words[i] = br.readLine();
             int len = words[i].length();
-            long place = 1;
+            int place = 1;
             for (int j = len - 1; j >= 0; j--) {
                 int c = words[i].charAt(j) - 'A';
                 weight[c] += place;
